+++ noodles-csi/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 [#157]: https://github.com/zaeleus/noodles/issues/157
 
+### Changed
+
+  * csi/index/reference_sequence/bin: Increase the visibility of `Builder`.
+
 ## 0.14.0 - 2023-03-03
 
 ### Changed